requirements.txt
numpy==1.24.4
opencv-python==4.8.0.74
Pillow=10.0.0
scipy-1.10.1
